Set default melee weapon attack range
