+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/MeleeWeapon.kts
@@ -1,0 +1,25 @@
+package world.gregs.voidps.world.interact.entity.player.combat.melee
+
+import world.gregs.voidps.engine.entity.Registered
+import world.gregs.voidps.engine.entity.character.contain.ItemChanged
+import world.gregs.voidps.engine.entity.character.player.Player
+import world.gregs.voidps.engine.entity.item.EquipSlot
+import world.gregs.voidps.engine.entity.item.Item
+import world.gregs.voidps.engine.entity.item.equipped
+import world.gregs.voidps.engine.entity.set
+import world.gregs.voidps.engine.event.Priority
+import world.gregs.voidps.engine.event.on
+import world.gregs.voidps.world.interact.entity.combat.weapon
+
+on<Registered>({ it.equipped(EquipSlot.Weapon).isNotEmpty() }, Priority.LOW) { player: Player ->
+    updateWeapon(player, player.equipped(EquipSlot.Weapon))
+}
+
+on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && item.isNotEmpty() }, Priority.LOW) { player: Player ->
+    updateWeapon(player, item)
+}
+
+fun updateWeapon(player: Player, weapon: Item) {
+    player["attack_range"] = 1
+    player.weapon = weapon
+}
