+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/special/StaffOfLight.kts
@@ -2,14 +2,16 @@ package world.gregs.voidps.world.interact.entity.player.combat.melee.special
 
 import world.gregs.voidps.engine.client.Colour
 import world.gregs.voidps.engine.client.variable.VariableSet
-import world.gregs.voidps.engine.entity.*
+import world.gregs.voidps.engine.entity.EffectStop
 import world.gregs.voidps.engine.entity.character.contain.ItemChanged
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
 import world.gregs.voidps.engine.entity.character.update.visual.setGraphic
+import world.gregs.voidps.engine.entity.hasEffect
 import world.gregs.voidps.engine.entity.item.EquipSlot
 import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.entity.item.equipped
+import world.gregs.voidps.engine.entity.start
+import world.gregs.voidps.engine.entity.stop
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.network.encode.message
@@ -23,21 +25,8 @@ import kotlin.math.floor
 
 fun isStaffOfLight(item: Item?) = item != null && item.name.startsWith("staff_of_light")
 
-on<Registered>({ isStaffOfLight(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
-    updateWeapon(player, player.equipped(EquipSlot.Weapon))
-}
-
-on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isStaffOfLight(item) }) { player: Player ->
-    updateWeapon(player, item)
-}
-
 on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isStaffOfLight(oldItem) }) { player: Player ->
     player.stop("power_of_light")
-}
-
-fun updateWeapon(player: Player, weapon: Item) {
-    player["attack_range"] = 1
-    player.weapon = weapon
 }
 
 on<CombatSwing>({ !swung() && isStaffOfLight(it.weapon) }, Priority.LOW) { player: Player ->
