+++ game/src/main/kotlin/world/gregs/voidps/world/interact/entity/player/combat/melee/weapon/VeracsFlail.kts
@@ -1,13 +1,8 @@
 package world.gregs.voidps.world.interact.entity.player.combat.melee.weapon
 
-import world.gregs.voidps.engine.entity.Registered
-import world.gregs.voidps.engine.entity.character.contain.ItemChanged
 import world.gregs.voidps.engine.entity.character.player.Player
 import world.gregs.voidps.engine.entity.character.update.visual.setAnimation
-import world.gregs.voidps.engine.entity.item.EquipSlot
 import world.gregs.voidps.engine.entity.item.Item
-import world.gregs.voidps.engine.entity.item.equipped
-import world.gregs.voidps.engine.entity.set
 import world.gregs.voidps.engine.event.Priority
 import world.gregs.voidps.engine.event.on
 import world.gregs.voidps.world.interact.entity.combat.CombatHit
@@ -16,19 +11,6 @@ import world.gregs.voidps.world.interact.entity.combat.hit
 import world.gregs.voidps.world.interact.entity.combat.weapon
 
 fun isFlail(item: Item?) = item != null && item.name.startsWith("veracs_flail")
-
-on<Registered>({ isFlail(it.equipped(EquipSlot.Weapon)) }) { player: Player ->
-    updateWeapon(player, player.equipped(EquipSlot.Weapon))
-}
-
-on<ItemChanged>({ container == "worn_equipment" && index == EquipSlot.Weapon.index && isFlail(item) }) { player: Player ->
-    updateWeapon(player, item)
-}
-
-fun updateWeapon(player: Player, weapon: Item) {
-    player["attack_range"] = 1
-    player.weapon = weapon
-}
 
 on<CombatSwing>({ !swung() && isFlail(it.weapon) }, Priority.LOW) { player: Player ->
     player.setAnimation("veracs_flail_attack")
